requirements.txt
easyocr
pillow
opencv-python
google-generativeai
python-dotenv